Dark Mode Progress Begins
Added in a Dark Mode Option to popup.html, currently it's overridng a
handful of polymer styles to acheive a dark theme. It's stil a WIP, but
currently works. To activate it, it's kind of weird:

Click the extension icon in your browser bar, enable the Dark Mode and
then hit Save Settings. Currently, to revert back to normal, you must
uncheck dark mode, hit settings, and then refresh.