Stable improvements to DarkMode Theme
The way I was building the dark theme compatibility has been reworked.
Instead of toggling everything one by one in JS, it’s much easier to
assign a class to the body tag, and just override everything down the
line from there. darkMode.js and the respective vanilla.js are much
smaller and simply act to toggle the theme on and off.

More general improvements to the darkMode css has been made, as well as
keeping the theme across multiple conversations once it’s already
loaded into the DOM.