Dark Mode functional
Dark Mode is functional and working now. It can be enabled with the
extension popup and your setting will be remembered across logged-in
Chrome devices. background.js contains a script which fires on dom
creation to check chrome storage if darkmode is enabled or not. More
darkmode classes were added to the style.css.